Reduce domain knowledge from design library
- it should be closer to Android/design system than it is to the domain
