+++ design_library/build.gradle
@@ -5,7 +5,7 @@ apply plugin: 'kotlin-android-extensions'
 dependencies {
     // Not sure if this is a bug or expected, but anyone using design_library needs constraintLayout also, hence `api`
     // e: Supertypes of the following classes cannot be resolved. Please make sure you have the required dependencies in the classpath:
-    // class com.muvi.design_library.FilmSummaryView, unresolved supertypes: androidx.constraintlayout.widget.ConstraintLayout
+    // class com.muvi.design_library.ListItem1View, unresolved supertypes: androidx.constraintlayout.widget.ConstraintLayout
     api libraries.androidxConstraintLayout
 
     implementation project(':resources')
