+++ design_library/src/main/res/layout/item_film_summary.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.muvi.design_library.FilmSummaryView xmlns:android="http://schemas.android.com/apk/res/android"
+<com.muvi.design_library.ListItem1View xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="wrap_content" />
