+++ feed/src/main/java/com/muvi/feed/FilmSummaryAdapter.kt
@@ -6,7 +6,8 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
-import com.muvi.design_library.FilmSummaryView
+import com.muvi.design_library.ListItem1View
+import com.muvi.design_library.commaAmpersandList
 
 internal class FilmSummaryAdapter : ListAdapter<UiModel, FilmSummaryAdapter.ViewHolder>(Differ) {
 
@@ -16,14 +17,19 @@ internal class FilmSummaryAdapter : ListAdapter<UiModel, FilmSummaryAdapter.View
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         val uiModel = getItem(position)
-        val model = FilmSummaryView.Model(
-                title = uiModel.title,
-                year = uiModel.year,
+        val model = ListItem1View.Model(
                 poster = uiModel.poster,
-                directors = uiModel.directors,
+                title = title(uiModel.title, uiModel.year, uiModel.directors),
                 onClick = uiModel.onClick
         )
-        (holder.itemView as FilmSummaryView).bind(model)
+        (holder.itemView as ListItem1View).bind(model)
+    }
+
+    private fun title(title: String, year: CharSequence?, directors: List<String>): CharSequence {
+        val titleBuilder = StringBuilder(title)
+        year?.let { titleBuilder.append(" ($it)") }
+        directors.commaAmpersandList()?.let { titleBuilder.append(" $it") }
+        return titleBuilder.toString()
     }
 
     object Differ : DiffUtil.ItemCallback<UiModel>() {
